refactor(benchmark): move benchmark to a foreground service
OPPO ColorOS aggressively freezes background activities and even
force-releases PARTIAL_WAKE_LOCKs held by plain Activities (visible as
'add wakelock … to ForceReleaseWakeLock list' in OplusProxyWakeLock).
Foreground services with a sticky notification are respected — once the
app is also whitelisted in Settings → Battery → App Battery Management
("Allow background activity"), the benchmark runs cleanly with the
screen off and the device locked.

Architecture:
- BenchmarkForegroundService (NEW) — holds the wake lock, posts a
  sticky progress notification, and runs the entire benchmark loop.
  Uses Dispatchers.Default so it isn't tied to a UI thread. Stops
  itself when done; the OS reclaims the process.
- BenchmarkActivity — reduced from ~470 lines to ~60. Now a thin
  launcher: receives `am start` Intent extras, forwards them to the
  service via startForegroundService(), and finishes immediately.
  Existing Python wrapper (benchmark_latency.py) is unchanged — it
  still launches the Activity and reads progress from logcat.
- AndroidManifest — registers the new service with
  foregroundServiceType="dataSync" (reuses the existing
  FOREGROUND_SERVICE_DATA_SYNC permission) and android:process=
  ":benchmark" (same isolated process as BenchmarkActivity).

Verified end-to-end on OPPO Snapdragon 8 Elite (OPD2413): launched the
benchmark, locked the screen 30 s in, watched a full medium_01 RAG run
complete with `hans_freeze=0` Athena freeze events and the screen
asleep through the entire decode phase. TTFT 1025 ms, total 13.8 s —
within the same envelope as foreground runs.

Pre-flight on a new device:
  Settings → Battery → App Battery Management → MAM-AI
                                              → "Allow background activity"

Without that, OPPO's OplusHansManager freezes the process at the OS
level regardless of foreground-service status.
